Try to change travis build ordering to fix compiletests
Multiple versions of the maud base crate get built, so compiletest
doesn't know which one to use in the same way that cargo would.
A possible solution is to build a canonical version of the crate first
(no features) and then explicitly run the compiletest tests. After, we
run builds and tests as normal, excluding the compiletest tests.