[actix-web] Add support for actix-web >= 2.0.0
In order to support both actix-web < 2.0.0 and >= 2.0.0 at the same
time I added a feature actix-web-2, which enables the actix-web
dependency with the respective version und the name active-web-2-dep
and the futures dependency which is required(?) for implementing the
Responder trait for >= 2.0.0.

Thus the user of maud has the choice to use actix-web in the version
they prefer.